% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htrlnorm.R
\name{htrlnorm}
\alias{htrlnorm}
\alias{mle.htrlnorm,dhtrlnorm}
\alias{mle.htrlnorm}
\alias{dhtrlnorm}
\alias{qhtrlnorm}
\alias{rhtrlnorm}
\title{Hurdle Truncated Log-Normal Model}
\usage{
mle.htrlnorm(x, warning.silent = TRUE)

dhtrlnorm(xi, phi = 0, meanlog = 0, sdlog = 1)

qhtrlnorm(p, phi = 0, meanlog = 0, sdlog = 1)

rhtrlnorm(N, phi = 0, meanlog = 0, sdlog = 1)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{warning.silent}{logical; if TRUE suppress all warning messages from
\code{\link{fitdistr}} during the mle process.}

\item{meanlog}{mean in the log scale.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{sd}{standard deviation in log scale.}
}
\description{
mle parameter estimation, random generation, density and
quantile function for the hurdle truncated log-normal distribution.
}
