% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htrlnorm.R
\name{htrlnorm}
\alias{htrlnorm}
\alias{mle.htrlnorm,dhtrlnorm,phtrlnorm,qhtrlnorm}
\alias{mle.htrlnorm}
\alias{dhtrlnorm}
\alias{qhtrlnorm}
\alias{rhtrlnorm}
\title{Hurdle Truncated Log-Normal Model}
\usage{
mle.htrlnorm(y, warning.silent = TRUE)

dhtrlnorm(x, phi = 0, meanlog = 0, sdlog = 1, a = 0, b = Inf)

qhtrlnorm(p, phi = 0, meanlog = 0, sdlog = 1, a = 0, b = Inf)

rhtrlnorm(n, phi = 0, meanlog = 0, sdlog = 1, a = 0, b = Inf)
}
\arguments{
\item{y}{numeric vector containing only finite values.}

\item{warning.silent}{logical; if TRUE suppress all warning messages from
\code{\link{fitdistr}} during the mle process.}

\item{x}{vector of quantiles.}

\item{phi}{probability of zero.}

\item{meanlog}{mean in the log scale.}

\item{sdlog}{standard deviation in log scale.}

\item{a}{vector of lower bounds.}

\item{b}{vector of upper bounds}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
mle parameter estimation, random generation, density and
quantile function for the hurdle truncated log-normal distribution.
}
